added some more content and images
I do completely disagree with the idea of having commit messages in the present tense, except of course for things like 'fixes #307'.
Anyway, now we have some content and a little bit of styling. It did kinda drive me crazy to not be able to use bootstrap from the get-go in this course. I mean, not that I really enjoy front-end work. It's okay if you're your own client, otherwise...rrgh.